+++ libs/ui-base-components/src/components/conditional-link/conditional-link.tsx
@@ -1,0 +1,12 @@
+export type ConditionalLinkProps = {
+  children: React.ReactNode
+} & React.AnchorHTMLAttributes<HTMLAnchorElement>
+
+export function ConditionalLink({ children, href, target, ...rest }: ConditionalLinkProps) {
+  if (!href) return children
+  return (
+    <a href={href} rel={target === '_blank' ? 'noreferrer' : undefined} {...rest}>
+      {children}
+    </a>
+  )
+}