+++ libs/ui-card/src/components/card/card.stories.tsx
@@ -1,6 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react'
 import { Card } from './'
 import { CardProps } from './'
+import { ConditionalLink } from '@r7/ui-base-components'
 
 const meta: Meta<CardProps> = {
   title: 'ui-card/Card',
@@ -16,22 +17,22 @@ type Story = StoryObj<CardProps>
 export const Primary: Story = {
   render: ({ ...args }) => (
     <Card {...args}>
-      <Card.HatWrapper>
-        <Card.HatImage
-          imageSource="//img.r7.com/images/concurso-publico-14032022123440824?dimensions=128x128"
-          description="Human hand writting in a paper"
-        />
-        <Card.HatTitle>News section</Card.HatTitle>
-      </Card.HatWrapper>
-      <Card.Title as="h2" fontStyle="heading-level-1">
-        The McRib is back (again): How a McNugget shortage led to its rise
-      </Card.Title>
+      <ConditionalLink href="https://www.google.com" title="Google">
+        <Card.HatWrapper>
+          <Card.HatImage
+            imageSource="//img.r7.com/images/concurso-publico-14032022123440824?dimensions=128x128"
+            description="Human hand writting in a paper"
+          />
+          <Card.HatTitle>News section</Card.HatTitle>
+        </Card.HatWrapper>
+        <Card.Title as="h2" fontStyle="heading-level-1">
+          The McRib is back (again): How a McNugget shortage led to its rise
+        </Card.Title>
+      </ConditionalLink>
     </Card>
   ),
   args: {
-    newsUrlTitle: 'The McRib is back (again): How a McNugget shortage led to its rise',
-    newsUrl: 'https://www.google.com',
-    openInBlank: true,
+    className: 'card-flex-col',
   },
   parameters: {
     design: {
@@ -65,10 +66,7 @@ export const ImageToTheSides: Story = {
     </Card>
   ),
   args: {
-    className: 'flex',
-    newsUrlTitle: 'The McRib is back (again): How a McNugget shortage led to its rise',
-    newsUrl: 'https://www.google.com',
-    openInBlank: true,
+    className: 'card-flex',
   },
   parameters: {
     design: {
@@ -103,9 +101,7 @@ export const ImageAbove: Story = {
     </Card>
   ),
   args: {
-    newsUrlTitle: 'The McRib is back (again): How a McNugget shortage led to its rise',
-    newsUrl: 'https://www.google.com',
-    openInBlank: true,
+    className: 'card-flex-col',
   },
   parameters: {
     design: {
@@ -119,7 +115,7 @@ export const ImageAbove: Story = {
 export const TitleOverImage: Story = {
   render: ({ ...args }) => (
     <Card {...args}>
-      <Card.Image format="landscape" shadow>
+      <Card.Image className="card-w-full" format="landscape" shadow>
         <img
           src="//img.r7.com/images/pantano-australia-rosa-brilhante-04102023182425856?resize=536x326&crop=691x420 80 0&dimensions=536x326"
           alt="cat"
@@ -142,10 +138,7 @@ export const TitleOverImage: Story = {
     </Card>
   ),
   args: {
-    className: 'relative',
-    newsUrlTitle: 'The McRib is back (again): How a McNugget shortage led to its rise',
-    newsUrl: 'https://www.google.com',
-    openInBlank: true,
+    className: 'card-relative',
   },
   parameters: {
     design: {