+++ libs/ui-base-components/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [0.3.0](https://gitlab.ir7.com.br/r7/front-monorepo/compare/ui-base-components-0.2.0...ui-base-components-0.3.0) (2023-11-27)
+
+### Features
+
+- **base-components:** adding dangerouslyHTML ([1ed0dd3](https://gitlab.ir7.com.br/r7/front-monorepo/commit/1ed0dd380d93562d6691671cfe699d13c397231f))
+- new conditionalLink component in ui-base-componnents; updated tests and stories of ui-card ([b21fc95](https://gitlab.ir7.com.br/r7/front-monorepo/commit/b21fc95bb1ad1146714dafb7fa6eb763a853a738))
+
 # [0.2.0](https://gitlab.ir7.com.br/r7/front-monorepo/compare/ui-base-components-0.1.0...ui-base-components-0.2.0) (2023-11-17)
 
 # 0.1.0 (2023-10-31)